Create BlockLight base class
