+++ src/main/java/dev/theatricalmod/theatrical/block/light/BlockGenericFixture.java
@@ -2,26 +2,20 @@ package dev.theatricalmod.theatrical.block.light;
 
 import dev.theatricalmod.theatrical.api.capabilities.power.TheatricalPower;
 import dev.theatricalmod.theatrical.api.fixtures.Fixture;
-import dev.theatricalmod.theatrical.block.BlockHangable;
 import dev.theatricalmod.theatrical.block.TheatricalBlocks;
 import dev.theatricalmod.theatrical.compat.top.ITOPInfoProvider;
 import dev.theatricalmod.theatrical.entity.FallingLightEntity;
 import dev.theatricalmod.theatrical.items.ItemPositioner;
-import dev.theatricalmod.theatrical.tiles.lights.TileEntityFixture;
 import dev.theatricalmod.theatrical.tiles.lights.TileEntityGenericFixture;
 import mcjty.theoneprobe.api.IProbeHitData;
 import mcjty.theoneprobe.api.IProbeInfo;
 import mcjty.theoneprobe.api.ProbeMode;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockRenderType;
 import net.minecraft.block.BlockState;
-import net.minecraft.entity.item.FallingBlockEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.ServerPlayerEntity;
 import net.minecraft.inventory.container.INamedContainerProvider;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTUtil;
-import net.minecraft.state.StateContainer.Builder;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.ActionResultType;
 import net.minecraft.util.Direction;
@@ -29,41 +23,14 @@ import net.minecraft.util.Hand;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.BlockRayTraceResult;
 import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.world.IBlockReader;
 import net.minecraft.world.IWorld;
 import net.minecraft.world.World;
-import net.minecraft.world.server.ServerWorld;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.fml.network.NetworkHooks;
 
-import javax.annotation.Nullable;
-import java.util.Random;
-
-public class BlockGenericFixture extends BlockHangable implements ITOPInfoProvider {
-
-    private final Fixture fixture;
+public class BlockGenericFixture extends BlockLight implements ITOPInfoProvider {
 
     public BlockGenericFixture(Fixture fixture) {
-        super(TheatricalBlocks.LIGHT_PROPERTIES);
-        this.fixture = fixture;
-    }
-
-    @Override
-    public boolean hasTileEntity(BlockState state) {
-        return true;
-    }
-
-    @Nullable
-    @Override
-    public TileEntity createTileEntity(BlockState state, IBlockReader world) {
-        TileEntityFixture tileEntityFixture = this.fixture.getFixtureType().getTileClass().get();
-        tileEntityFixture.setFixture(fixture);
-        return tileEntityFixture;
-    }
-
-    public Fixture getFixture() {
-        return fixture;
+        super(TheatricalBlocks.LIGHT_PROPERTIES, fixture);
     }
 
     @Override
@@ -74,22 +41,6 @@ public class BlockGenericFixture extends BlockHangable implements ITOPInfoProvid
             //N.B. Block removal is handled in the first tick of the entity because...reasons (vanilla does it)
         }
         return super.updatePostPlacement(state, from, fromState, world, pos, fromPos);
-    }
-
-    @OnlyIn(Dist.CLIENT)
-    @Override
-    public float getAmbientOcclusionLightValue(BlockState state, IBlockReader worldIn, BlockPos pos) {
-        return 1.0F;
-    }
-
-    @Override
-    public boolean propagatesSkylightDown(BlockState state, IBlockReader reader, BlockPos pos) {
-        return true;
-    }
-
-    @Override
-    protected void fillStateContainer(Builder<Block, BlockState> builder) {
-        super.fillStateContainer(builder);
     }
 
     @Override
@@ -123,9 +74,7 @@ public class BlockGenericFixture extends BlockHangable implements ITOPInfoProvid
 
         if (tileEntity instanceof TileEntityGenericFixture) {
             TileEntityGenericFixture pipe = (TileEntityGenericFixture) tileEntity;
-            pipe.getCapability(TheatricalPower.CAP, null).ifPresent(iTheatricalPowerStorage -> {
-                iProbeInfo.text(new StringTextComponent("Power: " + iTheatricalPowerStorage.getEnergyStored()));
-            });
+            pipe.getCapability(TheatricalPower.CAP, null).ifPresent(iTheatricalPowerStorage -> iProbeInfo.text(new StringTextComponent("Power: " + iTheatricalPowerStorage.getEnergyStored())));
         }
     }
 }
