+++ src/main/java/dev/theatricalmod/theatrical/block/light/BlockIntelligentFixture.java
@@ -1,14 +1,9 @@
 package dev.theatricalmod.theatrical.block.light;
 
 import dev.theatricalmod.theatrical.api.fixtures.Fixture;
-import dev.theatricalmod.theatrical.block.BlockHangable;
-import dev.theatricalmod.theatrical.tiles.lights.TileEntityFixture;
-import javax.annotation.Nullable;
 import net.minecraft.block.Block;
-import net.minecraft.block.BlockRenderType;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
-import net.minecraft.entity.EntityType;
 import net.minecraft.entity.item.FallingBlockEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.ServerPlayerEntity;
@@ -23,61 +18,16 @@ import net.minecraft.util.Direction;
 import net.minecraft.util.Hand;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.BlockRayTraceResult;
-import net.minecraft.world.IBlockReader;
 import net.minecraft.world.World;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.fml.network.NetworkHooks;
 
-public class BlockIntelligentFixture extends BlockHangable {
+public class BlockIntelligentFixture extends BlockLight {
 
     public static final BooleanProperty HANGING = BlockStateProperties.HANGING;
-    private final Fixture fixture;
 
     public BlockIntelligentFixture(Fixture fixture, Properties builder) {
-        super(builder);
-        this.fixture = fixture;
+        super(builder, fixture);
     }
-
-    @Override
-    public boolean hasTileEntity(BlockState state) {
-        return true;
-    }
-
-    @Nullable
-    @Override
-    public TileEntity createTileEntity(BlockState state, IBlockReader world) {
-        TileEntityFixture tileEntityFixture = this.fixture.getFixtureType().getTileClass().get();
-        tileEntityFixture.setFixture(fixture);
-        return tileEntityFixture;
-    }
-
-    @Override
-    public BlockRenderType getRenderType(BlockState state) {
-        return BlockRenderType.ENTITYBLOCK_ANIMATED;
-    }
-
-    @OnlyIn(Dist.CLIENT)
-    public float getAmbientOcclusionLightValue(BlockState state, IBlockReader worldIn, BlockPos pos) {
-        return 1.0F;
-    }
-
-    public boolean propagatesSkylightDown(BlockState state, IBlockReader reader, BlockPos pos) {
-        return true;
-    }
-
-    public boolean causesSuffocation(BlockState state, IBlockReader worldIn, BlockPos pos) {
-        return false;
-    }
-
-    public boolean isNormalCube(BlockState state, IBlockReader worldIn, BlockPos pos) {
-        return false;
-    }
-
-    public boolean canEntitySpawn(BlockState state, IBlockReader worldIn, BlockPos pos, EntityType<?> type) {
-        return false;
-    }
-
 
     @Override
     public BlockState getStateForPlacement(BlockItemUseContext context) {
