+++ src/main/java/dev/theatricalmod/theatrical/block/light/BlockLight.java
@@ -1,0 +1,57 @@
+package dev.theatricalmod.theatrical.block.light;
+
+import dev.theatricalmod.theatrical.api.fixtures.Fixture;
+import dev.theatricalmod.theatrical.block.BlockHangable;
+import dev.theatricalmod.theatrical.tiles.lights.TileEntityFixture;
+import net.minecraft.block.BlockRenderType;
+import net.minecraft.block.BlockState;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockReader;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+
+import javax.annotation.Nullable;
+
+public class BlockLight extends BlockHangable {
+
+    protected final Fixture fixture;
+
+    protected BlockLight(Properties builder, Fixture fixture) {
+        super(builder);
+        this.fixture = fixture;
+    }
+
+    @Override
+    public boolean hasTileEntity(BlockState state) {
+        return true;
+    }
+
+    @Nullable
+    @Override
+    public TileEntity createTileEntity(BlockState state, IBlockReader world) {
+        TileEntityFixture tileEntityFixture = this.fixture.getFixtureType().getTileClass().get();
+        tileEntityFixture.setFixture(fixture);
+        return tileEntityFixture;
+    }
+
+    @Override
+    public BlockRenderType getRenderType(BlockState state) {
+        return BlockRenderType.ENTITYBLOCK_ANIMATED;
+    }
+
+    @OnlyIn(Dist.CLIENT)
+    @Override
+    public float getAmbientOcclusionLightValue(BlockState state, IBlockReader worldIn, BlockPos pos) {
+        return 1.0F;
+    }
+
+    @Override
+    public boolean propagatesSkylightDown(BlockState state, IBlockReader reader, BlockPos pos) {
+        return true;
+    }
+
+    public Fixture getFixture() {
+        return fixture;
+    }
+}
