+++ src/main/java/dev/theatricalmod/theatrical/block/light/BlockMovingLight.java
@@ -3,17 +3,11 @@ package dev.theatricalmod.theatrical.block.light;
 import dev.theatricalmod.theatrical.block.TheatricalBlocks;
 import dev.theatricalmod.theatrical.fixtures.TheatricalFixtures;
 import dev.theatricalmod.theatrical.tiles.lights.TileEntityIntelligentFixture;
-import javax.annotation.Nullable;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockRenderType;
 import net.minecraft.block.BlockState;
-import net.minecraft.entity.EntityType;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockReader;
-import net.minecraft.world.World;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
+
+import javax.annotation.Nullable;
 
 public class BlockMovingLight extends BlockIntelligentFixture {
 
@@ -32,15 +26,5 @@ public class BlockMovingLight extends BlockIntelligentFixture {
         TileEntityIntelligentFixture tileEntityIntelligentFixture =  new TileEntityIntelligentFixture();
         tileEntityIntelligentFixture.setFixture(TheatricalFixtures.MOVING_LIGHT);
         return tileEntityIntelligentFixture;
-    }
-
-    @Override
-    public BlockRenderType getRenderType(BlockState state) {
-        return BlockRenderType.ENTITYBLOCK_ANIMATED;
-    }
-
-    @OnlyIn(Dist.CLIENT)
-    public float getAmbientOcclusionLightValue(BlockState state, IBlockReader worldIn, BlockPos pos) {
-        return 1.0F;
     }
 }
